requirements.txt
ollama>=0.1.6
google-search-results>=2.4.2
python-dotenv>=1.0.0
gradio>=4.12.0
requests>=2.31.0 